refactor: extract upgrade management and db initialization to UpgradeManagementRepository
